server/requirements.txt
grpcio==1.62.0
grpcio-tools==1.62.0
protobuf==4.24.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
